TAP-2630-feat(engine):Implement model deduction capabilities on the engine side.

+++ manager/tm/src/main/java/com/tapdata/tm/metadatainstance/service/MetadataInstancesServiceImpl.java
@@ -972,7 +972,7 @@ public class MetadataInstancesServiceImpl extends MetadataInstancesService{
 
 
     public int bulkSave(List<MetadataInstancesDto> insertMetaDataDtos,
-                        Map<String, MetadataInstancesDto> updateMetaMap, UserDetail userDetail, boolean saveHistory, String taskId, String uuid) {
+                        Map<String, MetadataInstancesDto> updateMetaMap, UserDetail userDetail, boolean saveHistory, String taskId, String uuid,Boolean isLastBatch) {
 
         BulkOperations bulkOperations = repository.bulkOperations(BulkOperations.BulkMode.UNORDERED);
 
@@ -1116,7 +1116,7 @@ public class MetadataInstancesServiceImpl extends MetadataInstancesService{
                 qualifiedNameLinkLogic(qualifiedNames, userDetail, taskId);
             }
 
-            if (StringUtils.isNotBlank(uuid) && !saveHistory) {
+            if (StringUtils.isNotBlank(uuid) && !saveHistory && isLastBatch) {
                 Criteria deleteOldMetadata = Criteria.where("taskId").is(taskId)
                         .and("transformUuid").ne(uuid);
                 deleteAll(new Query(deleteOldMetadata), userDetail);
